% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{api_get_key}
\alias{api_get_key}
\title{Get the OpenAI API key from environment variable}
\usage{
api_get_key()
}
\value{
API key string
}
\description{
Get the OpenAI API key from environment variable
}
\seealso{
[api_set_key()]
}
