% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{api_set_key}
\alias{api_set_key}
\title{Store the OpenAI API key as environment variable}
\usage{
api_set_key(api_key)
}
\arguments{
\item{api_key}{string, OpenAI API key
(see https://platform.openai.com/account/api-keys)}
}
\value{
api_key
}
\description{
Store the OpenAI API key as environment variable
}
\examples{
\dontrun{
api_set_key("my-secret-api-key-string")
}

}
\seealso{
[api_get_key()]
}
