% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{api_upload_file}
\alias{api_upload_file}
\title{Create "uploaded file" object}
\usage{
api_upload_file(f, type = NULL)
}
\arguments{
\item{f}{string/raw, content of file or path to the file}

\item{type}{mime type of path. If not supplied, will be guess by
\code{\link[mime:guess_type]{mime::guess_type()}} when needed.}
}
\value{
NULL if `f` was NULL otherwise "uploaded_file" object
}
\description{
See \link[httr]{upload_file}
}
