% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_assistants.R
\name{assistants_list_asistants_request}
\alias{assistants_list_asistants_request}
\title{API assistants: list assistants}
\usage{
assistants_list_asistants_request(
  assistant_id,
  limit = NULL,
  order = NULL,
  after = NULL,
  before = NULL,
  api_key = api_get_key()
)

assistants_list_asistants_request(
  assistant_id,
  limit = NULL,
  order = NULL,
  after = NULL,
  before = NULL,
  api_key = api_get_key()
)
}
\arguments{
\item{assistant_id}{string, the ID of the assistant the file belongs to.}

\item{limit}{NULL/integer, a limit on the number of objects to be returned. Limit can range between 1 and 100,
and the default is 20.}

\item{order}{NULL/string, sort order by the created_at timestamp of the objects. asc for ascending order and desc for
descending order. Defaults to desc}

\item{after}{NULL/string, a cursor for use in pagination. after is an object ID that defines your place in the list.
For instance, if you make a list request and receive 100 objects, ending with obj_foo, your subsequent call can
include after=obj_foo in order to fetch the next page of the list.}

\item{before}{NULL/string, a cursor for use in pagination. before is an object ID that defines your place in the
list. For instance, if you make a list request and receive 100 objects, ending with obj_foo, your subsequent call can
include before=obj_foo in order to fetch the previous page of the list.}

\item{api_key}{string, OpenAI API key
(see https://platform.openai.com/account/api-keys)}
}
\value{
\link[httr]{content} of the httr \link[httr]{response} object
or SimpleError (\link[base]{conditions}) enhanced with
two additional fields: `status_code` (response$status_code)
and `message_long` (built on response content)
}
\description{
Returns a list of assistants. To get more details, visit
https://platform.openai.com/docs/api-reference/assistants/listAssistants
https://platform.openai.com/docs/assistants

Returns a list of assistant files. To get more details, visit
https://platform.openai.com/docs/api-reference/assistants/listAssistantFiles
https://platform.openai.com/docs/assistants
}
