% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_assistants.R
\name{assistants_retrieve_assistant_file_request}
\alias{assistants_retrieve_assistant_file_request}
\title{API assistants: retrieve assistant file}
\usage{
assistants_retrieve_assistant_file_request(
  assistant_id,
  file_id,
  api_key = api_get_key()
)
}
\arguments{
\item{assistant_id}{string, the ID of the assistant who the file belongs to.}

\item{file_id}{string, the ID of the file we're getting.}

\item{api_key}{string, OpenAI API key
(see https://platform.openai.com/account/api-keys)}
}
\value{
\link[httr]{content} of the httr \link[httr]{response} object
or SimpleError (\link[base]{conditions}) enhanced with
two additional fields: `status_code` (response$status_code)
and `message_long` (built on response content)
}
\description{
Retrieves an AssistantFile. To get more details, visit
https://platform.openai.com/docs/api-reference/assistants/getAssistantFile
https://platform.openai.com/docs/assistants
}
