% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_assistants.R
\name{assistants_retrieve_assistant_request}
\alias{assistants_retrieve_assistant_request}
\title{API assistants: retrieve assistant}
\usage{
assistants_retrieve_assistant_request(assistant_id, api_key = api_get_key())
}
\arguments{
\item{assistant_id}{string, the ID of the assistant to retrieve.}

\item{api_key}{string, OpenAI API key
(see https://platform.openai.com/account/api-keys)}
}
\value{
\link[httr]{content} of the httr \link[httr]{response} object
or SimpleError (\link[base]{conditions}) enhanced with
two additional fields: `status_code` (response$status_code)
and `message_long` (built on response content)
}
\description{
Retrieves an assistant. To get more details, visit
https://platform.openai.com/docs/api-reference/assistants/getAssistant
https://platform.openai.com/docs/assistants
}
