% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{browseable_audio}
\alias{browseable_audio}
\title{Create browseable HTML audio}
\usage{
browseable_audio(data, format = "mp3")
}
\arguments{
\item{data}{audio data}

\item{format}{audio format}
}
\value{
HTML audio
}
\description{
Create browseable HTML audio
}
