% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_chat.R
\name{chat_fetch_messages}
\alias{chat_fetch_messages}
\title{Fetch messages (list of chat message "objects") from response content}
\usage{
chat_fetch_messages(res_content)
}
\arguments{
\item{res_content}{response object returned by \link{chat_request}}
}
\value{
list of chat message "objects" (see \link{chat_message})
}
\description{
Fetch messages (list of chat message "objects") from response content
}
