% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_chat.R
\name{chat_message}
\alias{chat_message}
\title{Create chat message "object"}
\usage{
chat_message(content, role = "user")
}
\arguments{
\item{content}{string, message content}

\item{role}{string, message role ("owner")}
}
\value{
message "object"
}
\description{
Create chat message "object"
}
