% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_completions.R
\name{completions_fetch_text}
\alias{completions_fetch_text}
\title{Fetch completions text from response content as dialog data.frame}
\usage{
completions_fetch_text(res_content, role = "ai", ltrim = TRUE)
}
\arguments{
\item{res_content}{response object returned by \link{completions_request}}

\item{role}{string, dialog role (phrase owner)}

\item{ltrim}{flag, trim left whitespace characters}
}
\value{
dialog data.frame
}
\description{
Fetch completions text from response content as dialog data.frame
}
\seealso{
[dialog_df()]
}
