% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog.R
\name{csv_to_dialog_df}
\alias{csv_to_dialog_df}
\title{Read csv file as dialog data.frame}
\usage{
csv_to_dialog_df(datapath)
}
\arguments{
\item{datapath}{string, csv file path}
}
\value{
Content of the input csv file as dialog data.frame,
SimpleError when an error occurs
}
\description{
Read csv file as dialog data.frame
}
