% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_cli}
\alias{demo_cli}
\title{Run one of the console application}
\usage{
demo_cli(api_key = Sys.getenv("API_KEY"), show_source = TRUE)
}
\arguments{
\item{api_key}{string, OpenAI API key
(see https://platform.openai.com/account/api-keys)}

\item{show_source}{flag, should the code of the selected script be displayed
before executing it?}
}
\value{
unused
}
\description{
Run one of the console application from the package app/cli directory.
}
\examples{
\dontrun{
  demo_cli()
  demo_cli("super-secret-key")
}

}
