% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demoChatApp}
\alias{demoChatApp}
\title{Run demo chat console application}
\usage{
demoChatApp()
}
\value{
unused
}
\description{
Run simle demo chat application in the console. If you do not want to enter
the api key, set the API_KEY environment variable.
}
\examples{
\dontrun{
  demoChatApp()
}
}
