% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{df_col_dt_format}
\alias{df_col_dt_format}
\title{Replace unix timestamp column(s) to formated dt string}
\usage{
df_col_dt_format(df, col, dt_format = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{df}{data.frame, input data.frame}

\item{col}{character vector, column names of the df that will be modified}

\item{dt_format}{string, date time format}
}
\value{
modified data.frame
}
\description{
Replace unix timestamp column(s) to formated dt string
}
