% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{df_col_obj_implode}
\alias{df_col_obj_implode}
\title{Change to string nested lists in a given data.frame}
\usage{
df_col_obj_implode(
  df,
  col,
  obj_prop,
  nested = TRUE,
  cell_header = "",
  objs_glue = "----\\n",
  cell_footer = "",
  obj_header = "",
  props_glue = "\\n",
  obj_footer = "",
  prop_fmt = "\%s: \%s"
)
}
\arguments{
\item{df}{data.frame, input data.frame}

\item{col}{character vector, df column names containing objects}

\item{obj_prop}{character vector, object properties}

\item{nested}{flag, whether the rows of the columns contain multiple objects}

\item{cell_header}{string/NULL, cell header}

\item{objs_glue}{string, how to combine objects}

\item{cell_footer}{string/NULL, cell footer}

\item{obj_header}{string/NULL, object header}

\item{props_glue}{string, how to combine properties}

\item{obj_footer}{string/NULL, object footer}

\item{prop_fmt}{string, sprintf fmt parameter with two `%s` fields (property
name, value)}
}
\value{
modified data.frame
}
\description{
Change to string nested lists in a given data.frame
}
