% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{df_exclude_cols}
\alias{df_exclude_cols}
\title{Remove columns from data.frame}
\usage{
df_exclude_cols(df, cols)
}
\arguments{
\item{df}{data.frame, input data.frame}

\item{cols}{character vector, column name(s) to be deleted}
}
\value{
modified data.frame
}
\description{
Remove columns from data.frame
}
