% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{df_null_replace}
\alias{df_null_replace}
\title{Replace all NULL values in given data.frame}
\usage{
df_null_replace(df, replacement = "")
}
\arguments{
\item{df}{data.frame, input data.frame}

\item{replacement}{string, replacement for NULL}
}
\value{
Modified input data.frame
}
\description{
Replace all NULL values in given data.frame
}
