% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{df_roxygen_tpl}
\alias{df_roxygen_tpl}
\title{df roxygen template}
\usage{
df_roxygen_tpl(df, on_missing_col)
}
\arguments{
\item{df}{data.frame, input data.frame}

\item{on_missing_col}{string, behavior for missing column(s):
"warn" - log warning, "skip" - skip missing column(s), "stop" - throw error}
}
\value{
Modified input data.frame
}
\description{
df roxygen template
}
\keyword{internal}
