% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog.R
\name{dialog_df}
\alias{dialog_df}
\title{Create dialog data.frame}
\usage{
dialog_df(content, role = "user", finish_reason = "stop")
}
\arguments{
\item{content}{string, message content}

\item{role}{string, message role ("owner")}

\item{finish_reason}{see \url{https://platform.openai.com/docs/guides/gpt/chat-completions-response-format}}
}
\value{
A one-row data.frame with columns: `content`, `role` and `finish_reason`
}
\description{
Create dialog data.frame
}
\examples{
dialog_df("some text message")
dialog_df("some another text message", role = "assistant")

}
