% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog.R
\name{dialog_df_to_csv}
\alias{dialog_df_to_csv}
\title{Save dialog data.frame as csv file}
\usage{
dialog_df_to_csv(dialog_df, file)
}
\arguments{
\item{dialog_df}{data.frame, dialog data.frame to save in csv file}

\item{file}{string, csv file path}
}
\value{
\link[utils]{write.table} return value or SimpleError
}
\description{
Save dialog data.frame as csv file
}
