% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_files.R
\name{files_fetch_list}
\alias{files_fetch_list}
\title{Extract files list as data.frame from response object}
\usage{
files_fetch_list(res_content)
}
\arguments{
\item{res_content}{response object returned by \link{files_list_request}}
}
\value{
Files list as data.frame
}
\description{
Extract files list as data.frame from response object
}
\examples{
\dontrun{
res_content <- files_list_request()
if (!is_error(res_content)) {
  files_list_df <- files_fetch_list(res_content)
  print(files_list_df)
}
}

}
