% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_files.R
\name{files_list_request}
\alias{files_list_request}
\title{API files: get list request}
\usage{
files_list_request(api_key)
}
\arguments{
\item{api_key}{string, API key}
}
\value{
content of the httr response object or SimpleError enhanced with
two additional fields: status_code (response$status_code) and message_long
(built on response content)
}
\description{
\url{https://platform.openai.com/docs/api-reference/files/list}
}
