% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_files.R
\name{files_retrieve_request}
\alias{files_retrieve_request}
\title{API files: retrieve request}
\usage{
files_retrieve_request(api_key, file_id)
}
\arguments{
\item{api_key}{string, API key}

\item{file_id}{string, id of the uploaded file}
}
\value{
content of the httr response object or SimpleError
}
\description{
\url{https://platform.openai.com/docs/api-reference/files/retrieve}
}
