% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fine-tunes.R
\name{fine_tunes_list_request}
\alias{fine_tunes_list_request}
\title{API fine tunes: list request}
\usage{
fine_tunes_list_request(api_key)
}
\arguments{
\item{api_key}{string, API key}
}
\value{
content of the httr response object or SimpleError enhanced with
two additional fields: status_code (response$status_code) and message_long
(built on response content)
}
\description{
\url{https://platform.openai.com/docs/api-reference/fine-tunes/list}
}
