% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fine-tunes.R
\name{fine_tunes_roxygen_tpl}
\alias{fine_tunes_roxygen_tpl}
\title{API fine tunes: roxygen template}
\usage{
fine_tunes_roxygen_tpl(api_key)
}
\arguments{
\item{api_key}{string, API key}
}
\value{
content of the httr response object or SimpleError enhanced with
two additional fields: status_code (response$status_code) and message_long
(built on response content)
}
\description{
API fine tunes: roxygen template
}
\keyword{internal}
