% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fine-tuning.R
\name{fine_tuning_fetch_jobs_list}
\alias{fine_tuning_fetch_jobs_list}
\title{API fine-tuning: extract fine-tuning jobs list from response object}
\usage{
fine_tuning_fetch_jobs_list(res_content)
}
\arguments{
\item{res_content}{response object returned by \link{fine_tuning_jobs_list_request}}
}
\value{
fine-tuning list models as data.frame
}
\description{
Extract fine-tuning jobs list as data.frame from response object
}
\examples{
\dontrun{
res_content <- fine_tuning_jobs_list_request()
if (!is_error(res_content)) {
  fine_tuning_jobs_df <- fine_tuning_fetch_jobs_list(res_content)
  print(fine_tuning_jobs_df)
}
}

}
