% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fine-tuning.R
\name{fine_tuning_fetch_retrived_job}
\alias{fine_tuning_fetch_retrived_job}
\title{API fine-tuning: fetch retrived job object from response object}
\usage{
fine_tuning_fetch_retrived_job(res_content)
}
\arguments{
\item{res_content}{response object returned by \link{fine_tuning_retrive_job_request}}
}
\value{
fine-tuning job object
}
\description{
Extract fine-tuning job object from response object
}
\examples{
\dontrun{
res_content <- fine_tuning_retrive_job_request("my-secret-api-key-string", "job-id")
if (!is_error(res_content)) {
  fine_tuning_events_df <- fine_tuning_fetch_events_list(res_content)
  print(fine_tuning_events_df)
}
}

}
