% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_browseable}
\alias{is_browseable}
\title{Test if RStudio Viewer (build in browser) is available}
\usage{
is_browseable()
}
\value{
TRUE/FALSE
}
\description{
Test if RStudio Viewer (build in browser) is available
}
