% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_images.R
\name{is_image_set}
\alias{is_image_set}
\title{Test if x is a image set}
\usage{
is_image_set(x)
}
\arguments{
\item{x}{R variable to test}
}
\value{
TRUE/FALSE
}
\description{
Test if x is a image set - a list consisting of three elements: data, prompt and size
}
