% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog.R
\name{merge_dialog_df}
\alias{merge_dialog_df}
\title{Merge multiple dialog data.frame}
\usage{
merge_dialog_df(...)
}
\arguments{
\item{...}{dialog data.frame or NULL}
}
\value{
data.frame containing all input dialogs
}
\description{
Merge multiple dialog data.frame
}
\examples{
d1 <- dialog_df("message 1")
d2 <- dialog_df("message 2")
print(
  merge_dialog_df(
    d1,
    merge_dialog_df(d1, d2),
    NULL,
    d2
  )
)

}
