% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_messages.R
\name{messages_retrieve_message_file_request}
\alias{messages_retrieve_message_file_request}
\title{API messages: retrieve message file}
\usage{
messages_retrieve_message_file_request(
  thread_id,
  message_id,
  file_id,
  api_key = api_get_key()
)
}
\arguments{
\item{thread_id}{string, the ID of the thread the messages belong to}

\item{message_id}{string, the ID of the message that the files belongs to}

\item{file_id}{string, the ID of the file being retrieved}

\item{api_key}{string, OpenAI API key
(see https://platform.openai.com/account/api-keys)}
}
\value{
\link[httr]{content} of the httr \link[httr]{response} object
or SimpleError (\link[base]{conditions}) enhanced with
two additional fields: `status_code` (response$status_code)
and `message_long` (built on response content)
}
\description{
Retrieve a message file. To get more details, visit
https://platform.openai.com/docs/api-reference/messages/getMessageFile
https://platform.openai.com/docs/assistants
}
