% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_models.R
\name{models_list_request}
\alias{models_list_request}
\title{API models: list request}
\usage{
models_list_request(api_key)
}
\arguments{
\item{api_key}{string, API key}
}
\value{
content of the httr response object or SimpleError
}
\description{
\url{https://platform.openai.com/docs/api-reference/models/list}
}
