% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_models.R
\name{models_retrieve_request}
\alias{models_retrieve_request}
\title{API models: retrieve model request}
\usage{
models_retrieve_request(api_key, model)
}
\arguments{
\item{api_key}{string, OpenAI API key
(see https://platform.openai.com/account/api-keys)}

\item{model}{string, the ID of the model to use for this request}
}
\value{
\link[httr]{content} of the httr \link[httr]{response} object
or SimpleError (\link[base]{conditions}) enhanced with
two additional fields: `status_code` (response$status_code)
and `message_long` (built on response content)
}
\description{
Retrieves a model instance, providing basic information about the model such as the owner and permissioning.
To get more details, visit: https://platform.openai.com/docs/models
https://platform.openai.com/docs/api-reference/models/list
}
