% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_moderations.R
\name{moderations_object_request}
\alias{moderations_object_request}
\title{API moderations: get moderation object}
\usage{
moderations_object_request(api_key, id, model)
}
\arguments{
\item{api_key}{string, OpenAI API key
(see https://platform.openai.com/account/api-keys)}

\item{id}{string, The unique identifier for the moderation request}

\item{model}{string, The model used to generate the moderation results}
}
\value{
\link[httr]{content} of the httr \link[httr]{response} object
or SimpleError (\link[base]{conditions}) enhanced with
two additional fields: `status_code` (response$status_code)
and `message_long` (built on response content)
}
\description{
Represents policy compliance report by OpenAI's content moderation model
against a given input. To get more details, visit
https://platform.openai.com/docs/api-reference/moderations/object
https://platform.openai.com/docs/guides/moderation
}
