% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fine-tuning.R
\name{print.oaii_fine_tuning_job}
\alias{print.oaii_fine_tuning_job}
\title{print S3 method for oaii_fine_tuning_job class}
\usage{
\method{print}{oaii_fine_tuning_job}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{print} prints its argument and returns it \emph{invisibly} (via
  \code{\link[base]{invisible}(x)}).  It is a generic function which means that
  new printing methods can be easily added for new \code{\link[base]{class}}es.
}
