% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.oaii_res_se}
\alias{print.oaii_res_se}
\title{Class oaii_res_se print S3 method}
\usage{
\method{print}{oaii_res_se}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{print} prints its argument and returns it \emph{invisibly} (via
  \code{\link[base]{invisible}(x)}).  It is a generic function which means that
  new printing methods can be easily added for new \code{\link[base]{class}}es.
}
