% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_runs.R
\name{runs_retrieve_run_request}
\alias{runs_retrieve_run_request}
\title{API runs: retrieve run}
\usage{
runs_retrieve_run_request(thread_id, run_id, api_key = api_get_key())
}
\arguments{
\item{thread_id}{string, The ID of the thread (https://platform.openai.com/docs/api-reference/threads) that was run}

\item{run_id}{string, the ID of the run to retrieve}

\item{api_key}{string, OpenAI API key
(see https://platform.openai.com/account/api-keys)}
}
\value{
\link[httr]{content} of the httr \link[httr]{response} object
or SimpleError (\link[base]{conditions}) enhanced with
two additional fields: `status_code` (response$status_code)
and `message_long` (built on response content)
}
\description{
Retrieves a thread. To get more details, visit
https://platform.openai.com/docs/api-reference/threads/getThread
https://platform.openai.com/docs/assistants
}
