% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{set_logger}
\alias{set_logger}
\title{Set log functions used by `oaii` package}
\usage{
set_logger(...)
}
\arguments{
\item{...}{parameters in form log_level = function}
}
\description{
Set log functions used by `oaii` package
}
\examples{
\dontrun{
logger <- log4r::logger("DEBUG")
log_error <- function(...) log4r::error(logger, ...)
log_warning <- function(...) log4r::warn(logger, ...)
log_info <- function(...) log4r::info(logger, ...)
log_debug <- function(...) log4r::debug(logger, ...)

oaii::set_logger(
  error = log_error,
  warning = log_warning,
  info = log_info,
  debug = log_debug
)
}

}
