% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_threads.R
\name{threads_retrieve_thread_request}
\alias{threads_retrieve_thread_request}
\title{API threads: retrieve thread}
\usage{
threads_retrieve_thread_request(thread_id, api_key = api_get_key())
}
\arguments{
\item{thread_id}{string, the ID of the thread to retrieve.}

\item{api_key}{string, OpenAI API key
(see https://platform.openai.com/account/api-keys)}
}
\value{
\link[httr]{content} of the httr \link[httr]{response} object
or SimpleError (\link[base]{conditions}) enhanced with
two additional fields: `status_code` (response$status_code)
and `message_long` (built on response content)
}
\description{
Retrieves a thread. To get more details, visit
https://platform.openai.com/docs/api-reference/threads/getThread
https://platform.openai.com/docs/assistants
}
