% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_chat.R
\name{chat_fetch_messages}
\alias{chat_fetch_messages}
\title{Fetch messages (dialog data.frame with chat messages) from response content}
\usage{
chat_fetch_messages(res_content)
}
\arguments{
\item{res_content}{response object returned by \link{chat_request}}
}
\value{
dialog data.frame (see \link{dialog_df})
}
\description{
Fetch messages (dialog data.frame with chat messages) from response content
}
\seealso{
[dialog_df()]
}
