% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_chat.R
\name{chat_fetch_messages}
\alias{chat_fetch_messages}
\title{Fetch messages from response content}
\usage{
chat_fetch_messages(res_content)
}
\arguments{
\item{res_content}{response object}
}
\description{
Fetch messages from response content
}
