% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_chat.R
\name{chat_merge_messages}
\alias{chat_merge_messages}
\title{Merge chat messages}
\usage{
chat_merge_messages(...)
}
\arguments{
\item{...}{chat message and/or messages objects}
}
\description{
Merge chat messages
}
