% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_chat.R
\name{chat_request}
\alias{chat_request}
\title{Send chat request
See \url{https://platform.openai.com/docs/api-reference/chat}}
\usage{
chat_request(
  api_key,
  messages,
  model = "gpt-3.5-turbo",
  temperature = 0.7,
  n = 1,
  max_tokens = 50,
  presence_penalty = 0,
  frequency_penalty = 0
)
}
\arguments{
\item{api_key}{API endpoint parameter}

\item{messages}{API endpoint parameter}

\item{model}{API endpoint parameter}

\item{temperature}{API endpoint parameter}

\item{n}{API endpoint parameter}

\item{max_tokens}{API endpoint parameter}

\item{presence_penalty}{API endpoint parameter}

\item{frequency_penalty}{API endpoint parameter}
}
\description{
Send chat request
See \url{https://platform.openai.com/docs/api-reference/chat}
}
