% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_completions.R
\name{completion_message}
\alias{completion_message}
\title{Create completion "message object"}
\usage{
completion_message(content, role = "user")
}
\arguments{
\item{content}{string, message content}

\item{role}{string, message "owner"}
}
\value{
completion "message object" - a list consisting of two elements: content and role
}
\description{
Create completion "message object"
}
