% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{df_order_by_col}
\alias{df_order_by_col}
\title{Sort data.frame by column name}
\usage{
df_order_by_col(df, col, decreasing = FALSE)
}
\arguments{
\item{df}{data.frame, input data.frame}

\item{col}{string, column name as sort source}

\item{decreasing}{flag, should the sort order be increasing or decreasing?}
}
\value{
modified data.frame
}
\description{
Sort data.frame by column name
}
