% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{df_roxygen_tpl}
\alias{df_roxygen_tpl}
\title{df roxygen template}
\usage{
df_roxygen_tpl(df)
}
\arguments{
\item{df}{data.frame, input data.frame}
}
\value{
modified data.frame
}
\description{
df roxygen template
}
\keyword{internal}
