% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_files.R
\name{files_delete_request}
\alias{files_delete_request}
\title{API files: delete request}
\usage{
files_delete_request(api_key, file_id)
}
\arguments{
\item{api_key}{string, API key}

\item{file_id}{string, id of the uploaded file}
}
\value{
content of the httr response object or SimpleError enhanced with
two additional fields: status_code (response$status_code) and message_long
(built on response content)
}
\description{
\url{https://platform.openai.com/docs/api-reference/files/delete}
}
