% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_completions.R, R/api_files.R,
%   R/api_models.R
\name{files_roxygen_tpl}
\alias{files_roxygen_tpl}
\title{API completions: roxygen template}
\usage{
files_roxygen_tpl(api_key)

files_roxygen_tpl(api_key)

files_roxygen_tpl(api_key)
}
\arguments{
\item{api_key}{string, API key}

\item{file_id}{string, id of the uploaded file}
}
\value{
content of the httr response object or SimpleError
}
\description{
API completions: roxygen template

API files: roxygen template

API models: roxygen template
}
\keyword{internal}
