% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_files.R
\name{files_upload_request}
\alias{files_upload_request}
\title{API files: upload request}
\usage{
files_upload_request(api_key, file, purpose)
}
\arguments{
\item{api_key}{string, OpenAI API key
(see https://platform.openai.com/account/api-keys)}

\item{file}{string, path or content of the JSON Lines file to be uploaded}

\item{purpose}{string, the intended purpose of the uploaded documents.
Use "fine-tune" for Fine-tuning.}
}
\value{
\link[httr]{content} of the httr \link[httr]{response} object
or SimpleError (\link[base]{conditions}) enhanced with
two additional fields: `status_code` (response$status_code)
and `message_long` (built on response content)
}
\description{
To get more details, visit https://platform.openai.com/docs/api-reference/files/upload
}
