% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_images.R
\name{images_generator_request}
\alias{images_generator_request}
\title{API images: create request}
\usage{
images_generator_request(
  api_key,
  prompt,
  response_format = NULL,
  size = NULL,
  n = NULL,
  user = NULL
)
}
\arguments{
\item{api_key}{string, API key}

\item{prompt}{string, a text description of the desired image(s).
The maximum length is 1000 characters.}

\item{response_format}{string,tThe format in which the generated images
are returned. Must be one of "url" or "b64_json".}

\item{size}{string, the size of the generated images. Must be one of
"256x256", "512x512" or "1024x1024"}

\item{n}{integer, the number of images to generate. Must be between 1 and 10.}

\item{user}{string a unique identifier representing your end-user,
which can help OpenAI to monitor and detect abuse.}
}
\value{
content of the httr response object or SimpleError enhanced with
two additional fields: status_code (response$status_code) and message_long
(built on response content)
}
\description{
\url{https://platform.openai.com/docs/api-reference/images/create}
}
