% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_images.R
\name{images_merge_sets}
\alias{images_merge_sets}
\title{Merge image set/sets}
\usage{
images_merge_sets(...)
}
\arguments{
\item{...}{images set(s), NULL also allowed}
}
\value{
List of image set(s)
}
\description{
Merge given image set/sets into single images sets object (list with image
sets). Have a look at \link{images_fech_set}.
}
