% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_models.R
\name{models_delete_request}
\alias{models_delete_request}
\title{API models: delete model request}
\usage{
models_delete_request(model, api_key = api_get_key())
}
\arguments{
\item{model}{string, the model to delete}

\item{api_key}{string, OpenAI API key
(see https://platform.openai.com/account/api-keys)}
}
\value{
\link[httr]{content} of the httr \link[httr]{response} object
or SimpleError (\link[base]{conditions}) enhanced with
two additional fields: `status_code` (response$status_code)
and `message_long` (built on response content)
}
\description{
Delete a fine-tuned model. You must have the Owner role in your organization to delete a model. To get more details,
visit https://platform.openai.com/docs/models
https://platform.openai.com/docs/api-reference/fine-tunes/delete-model
}
