% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_models.R
\name{models_list_request}
\alias{models_list_request}
\title{API models: list request}
\usage{
models_list_request(api_key)
}
\arguments{
\item{api_key}{string, OpenAI API key
(see \url{https://platform.openai.com/account/api-keys})}
}
\value{
\link[httr]{content} of the httr \link[httr]{response} object
or SimpleError (\link[base]{conditions}) enhanced with
two additional fields: `status_code` (response$status_code)
and `message_long` (built on response content)
}
\description{
\url{https://platform.openai.com/docs/api-reference/models/list}
}
\examples{
\dontrun{
res_content <- models_list_request("my-secret-api-key-string")
if (!is_error(res_content)) {
  models_list_df <- models_fetch_list(res_content)
  print(models_list_df)
}
}

}
