% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{request}
\alias{request}
\title{API request}
\usage{
request(endpoint, api_key, body = NULL, encode = "json", method = "POST")
}
\arguments{
\item{endpoint}{string, API endpoint}

\item{api_key}{string, API key}

\item{body}{NULL/list, request body}

\item{encode}{request encode}

\item{method}{request method}
}
\description{
API request
}
