% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{set_logger}
\alias{set_logger}
\title{Set log functions used by this package}
\usage{
set_logger(...)
}
\arguments{
\item{...}{parameters in form log_level = function}
}
\description{
Set log functions used by this package
}
